Remove bugged editor status report
When toggling status with a 'caret' selection the reported status of queryCommandState is sometimes false. Let the users using the editor handle this case manually and take responsibility for this part.

When modifying the most basic text possible, with a 'caret' selection inside a word, toggle B/I/U/S, write something and toggle I/U/S back. U and S should report the wrong state while I reports the correct state when calling queryCommandState on them with this bug