Detect missing execCommand notifications
This did happen for example in the case were you have a content of <h1>Hello Wolrd</h1>, and you select the word "hello" which will indicate that the state "bold" is enabled and the state "ilatic" is disabled. Now you enable "italic" and then you try disabling "bold". Bold will effectively be disabled but the "selectionchange" event listener won't pick up on it